+++ Tests/SwifterSwiftUITests/SwifterSwiftUI/SwiftUI/Modifiers/SizeViewModifier+Tests.swift
@@ -1,0 +1,8 @@
+//
+//  SizeViewModifier+Tests.swift
+//  SwifterSwiftUI
+//
+//  Created by Saul Moreno Abril on 12/06/2020.
+//
+
+import Foundation